extend-token finished

+++ src/frontend/sni/src/pages/Login.jsx
@@ -1,9 +1,13 @@
 import api from "../api/axios";
 import { useState } from "react";
-import { AxiosError } from "axios";
 
 const Login = () => {
   const [message, setMessage] = useState("");
+
+  if (api.isSignedIn()) {
+    window.location.href = '/dashboard'
+    return;
+  }
 
   const signIn = async () => {
     const username = document.getElementById("username").value;
@@ -11,18 +15,35 @@ const Login = () => {
 
     try {
       let data, status = {};
-      const response = await api.post(
+      let alreadySignedIn = false;
+
+      const response = await api.instance.post(
         `${import.meta.env.VITE_API_HOST}/auth/login`,
         JSON.stringify({ username, password })
       ).then((response) => {
         data = response.data;
         status = response.status;
+        // console.log(response.data.additional.token)
+
+        if (status === 202) {
+          api.setToken(response.data.additional.token)
+          console.log(api.getToken())
+          alreadySignedIn = true;
+        }
       })
         .catch((error) => {
-          data = error.response.data;
+          console.log(error);
+          // data = error.response.data;
           status = error.response.status;
         });
 
+      if (alreadySignedIn)
+      {
+        window.location.href = '/dashboard'
+        return;
+      }
+
+      localStorage.removeItem('jwt');
 
       if (status === 200) {
         localStorage.setItem("message", data.message);
@@ -32,27 +53,35 @@ const Login = () => {
         setMessage(data.message);
       }
     } catch (error) {
-      setMessage(error.response.data.message);
+      console.log(error);
+      setMessage("An unknown error happened.");
     }
   }
 
   return (
-    <div className="flex flex-col gap-4 items-start">
-      <h1>Hello!</h1>
-      <p className="text-2xl">Sign in to continue.</p>
+    <div className={"flex items-center justify-center h-screen"}>
+      <div className="flex flex-col gap-4 items-start">
+        <h1 className={"text-4xl"}>Hello!</h1>
+        <p className="text-2xl">Sign in to continue.</p>
 
-      <input id="username" type="text" placeholder="Username" className="w-80 h-12 p-4 rounded-xl text-lg" />
-      <input id="password" type="password" placeholder="Password" className="w-80 h-12 p-4 rounded-xl text-lg" />
+        <input id="username" type="text" placeholder="Username" className="w-80 h-12 p-4 rounded-xl text-lg" />
+        <input id="password" type="password" placeholder="Password" className="w-80 h-12 p-4 rounded-xl text-lg" />
 
-      <button className="w-full h-12 rounded-xl bg-neutral-700 text-white text-lg" onClick={async () => { await signIn(); }}>Sign in</button>
+        <button className="w-32 h-12 rounded-xl bg-neutral-700 text-white text-lg"
+                onClick={async () => { await signIn(); }}
+                onKeyDown={async (event) => { if (event.key === "Enter") { await signIn(); } }}
+        >
+          Sign in
+        </button>
 
-      <hr />
-      <p>{"Don't"} have an account? <a href="/register">Register here</a></p>
+        <hr />
+        <p>{"Don't"} have an account? <a href="/register">Register here</a></p>
 
-      {
-        message !== "" &&
-        <p className="text-lg text-red-400">{message}</p>
-      }
+        {
+          message !== "" &&
+          <p className="text-lg text-red-400">{message}</p>
+        }
+      </div>
     </div>
   )
 }
